+++ src/functions/drawSvgPath.ts
@@ -51,12 +51,11 @@ const quadraticBezierCurve = (points: number[][]) => {
   const first = points[0];
   let svgPath = `M${String(first[X])},${String(first[Y])}M`;
 
-  for (let i = 0; i < points.length; i++) {
-    const next = points[i];
+  for (const next of points) {
     const midPoint = getMidPoint(point[X], point[Y], next[X], next[Y]);
 
-    svgPath += ` ${midPoint[X]},${midPoint[Y]}`;
-    svgPath += `Q${next[X]},${next[Y]}`;
+    svgPath += ` ${String(midPoint[X])},${String(midPoint[Y])}`;
+    svgPath += `Q${String(next[X])},${String(next[Y])}`;
     point = next;
   }
